refactor(portal): promote shared state, server modules, and GraphQL docs to scenes/common
Move the cross-tree singletons — color-atom, the create-app and ban-dialog
atoms (split out; originals re-export), fetch-app-env helper+doc, the
remove-from-review hook/action/doc, fetch-team, delete-team + DeleteTeamDialog
server action, CreateAppDialog client/server docs + submit action + form
schema, getAccumulativeTransactionData, and the EnableWorldId40 self-managed
registration server helper — to the full-path common mirror. Atoms and write
paths must stay single-copy while v2 and v3 coexist. Consumers in both trees
repointed. No operation renames; no codegen run.
