+++ web/scenes/Portal/Profile/common/UserInfo/Icon/index.tsx
@@ -1,5 +1,5 @@
 import { TYPOGRAPHY, Typography } from "@/components/Typography";
-import { colorAtom } from "@/scenes/Portal/layout/color-atom";
+import { colorAtom } from "@/scenes/common/layout/color-atom";
 import clsx from "clsx";
 import { useAtom } from "jotai";
 import { CSSProperties, HTMLAttributes } from "react";
