+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Configuration/AppTopBar/index.tsx
@@ -8,7 +8,7 @@ import { Role_Enum } from "@/graphql/graphql";
 import { Button } from "@/components/Button";
 import { Auth0SessionUser } from "@/lib/types";
 import { getCDNImageUrl, getDefaultLogoImgCDNUrl } from "@/lib/utils";
-import { useRemoveFromReview } from "@/scenes/Portal/Teams/TeamId/Apps/common/hooks/use-remove-from-review";
+import { useRemoveFromReview } from "@/scenes/common/Teams/TeamId/Apps/common/hooks/use-remove-from-review";
 import { useUser } from "@auth0/nextjs-auth0/client";
 import clsx from "clsx";
 import { useAtom } from "jotai";
