+++ web/scenes/Portal/common/DeleteTeamDialog/index.tsx
@@ -17,7 +17,7 @@ import { useCallback, useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import { toast } from "react-toastify";
 import * as yup from "yup";
-import { deleteTeamServerSide } from "./server";
+import { deleteTeamServerSide } from "@/scenes/common/common/DeleteTeamDialog/server";
 
 type DeleteTeamDialogProps = DialogProps & {
   team: {
