+++ web/scenes/Portal/layout/CreateAppDialog/index-v4.tsx
@@ -27,9 +27,12 @@ import { EnableWorldId40Content } from "../../Teams/TeamId/Apps/AppId/EnableWorl
 import { SelfManagedTransactionInfoContent } from "../../Teams/TeamId/Apps/AppId/EnableWorldId40/SelfManagedTransactionInfo/SelfManagedTransactionInfoContent";
 import { GenerateNewKeyContent } from "../../Teams/TeamId/Apps/AppId/GenerateNewKey/GenerateNewKeyContent";
 import { UseExistingKeyContent } from "../../Teams/TeamId/Apps/AppId/UseExistingKey/UseExistingKeyContent";
-import { useRegisterRpMutation } from "./client/register-rp.generated";
-import { createAppSchemaV4, CreateAppSchemaV4 } from "./form-schema-v4";
-import { validateAndInsertAppServerSideV4 } from "./server/v4/submit";
+import { useRegisterRpMutation } from "@/scenes/common/layout/CreateAppDialog/client/register-rp.generated";
+import {
+  createAppSchemaV4,
+  CreateAppSchemaV4,
+} from "@/scenes/common/layout/CreateAppDialog/form-schema-v4";
+import { validateAndInsertAppServerSideV4 } from "@/scenes/common/layout/CreateAppDialog/server/v4/submit";
 
 type CreateDialogStep =
   | "create"
