+++ web/scenes/PortalV3/Teams/TeamId/Apps/AppId/page/AppStatsGraph/GraphsSection/use-get-accumulative-transactions.ts
@@ -3,7 +3,7 @@ import {
   getAccumulativePaymentsData,
   GetAccumulativePaymentsDataReturnType,
   GetAccumulativeTransactionsDataReturnType,
-} from "@/scenes/Portal/Teams/TeamId/Apps/AppId/MiniApp/Transactions/page/server/getAccumulativeTransactionData";
+} from "@/scenes/common/Teams/TeamId/Apps/AppId/MiniApp/Transactions/page/server/getAccumulativeTransactionData";
 
 export const useGetAccumulativeTransactions = (appId: string) => {
   const [payments, setPayments] =
