+++ web/scenes/common/Teams/TeamId/Apps/common/BanMessageDialog/atoms.ts
@@ -1,0 +1,3 @@
+import { atom } from "jotai";
+
+export const banMessageDialogOpenedAtom = atom<boolean>(false);
