+++ web/scenes/common/Teams/TeamId/Apps/common/hooks/graphql/server/remove-app-from-review.graphql
