+++ web/scenes/common/Teams/TeamId/Team/common/TeamProfile/graphql/client/fetch-team.generated.ts
