+++ web/scenes/common/common/DeleteTeamDialog/graphql/server/delete-team.graphql
