+++ web/scenes/common/common/DeleteTeamDialog/server/index.ts
