+++ web/scenes/common/layout/CreateAppDialog/client/register-rp.graphql
