+++ web/scenes/common/layout/CreateAppDialog/graphql/server/insert-app.generated.ts
