+++ web/scenes/common/layout/Header/atoms.ts
@@ -1,0 +1,3 @@
+import { atom } from "jotai";
+
+export const createAppDialogOpenedAtom = atom(false);
